Add .d.ts file for cms typings
